@@ -14,8 +14,7 @@
   "files": [
     "lib/**/*.js",
     "index.d.ts*",
-    "types",
-    "!lib/rules/test-utils.js"
+    "types"
   ],
   "scripts": {
     "build": "npx tsc",
@@ -28,23 +27,26 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.3",
-    "@eslint/js": "^9.39.2",
+    "@eslint/js": "^10.0.1",
     "@eslint/markdown": "^7.5.1",
-    "@stylistic/eslint-plugin": "^5.7.1",
-    "@types/node": "^25.2.1",
-    "eslint": "^9.39.2",
+    "@stylistic/eslint-plugin": "^5.8.0",
+    "@types/node": "^25.2.3",
+    "eslint": "^10.0.0",
     "eslint-plugin-eslint-plugin": "^7.3.0",
-    "eslint-plugin-jsdoc": "^62.5.2",
+    "eslint-plugin-jsdoc": "^62.5.4",
     "eslint-plugin-n": "^17.23.2",
     "markdownlint-cli": "^0.47.0",
-    "mocha": "^12.0.0-beta-6",
+    "mocha": "^12.0.0-beta-8",
     "nyc": "^17.1.0",
     "testdouble": "^3.20.2",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.54.0"
+    "typescript-eslint": "^8.55.0"
   },
   "peerDependencies": {
     "eslint": ">=7.7.0"
+  },
+  "overrides": {
+    "eslint": "^10.0.0"
   },
   "keywords": [
     "eslint",
