@@ -37,6 +37,7 @@ module.exports = {
      * @returns {SourceCode} the source-code object.
      */
     contextSourceCode: function(context) {
-        return context.sourceCode || context.getSourceCode();
+        return context.sourceCode
+            || /** @type {RuleContext & { getSourceCode: () => SourceCode }} */(context).getSourceCode();
     }
 };
