@@ -803,7 +803,7 @@ const headerRule = {
                                         },
                                         messageId: "incorrectHeaderLine",
                                         data: {
-                                            pattern: headerLine
+                                            pattern: headerLine.toString()
                                         },
                                         fix: canFix
                                             ? genReplaceFixer(
@@ -847,7 +847,7 @@ const headerRule = {
 
                 /** @type {null | string} */
                 let errorMessageId = null;
-                /** @type {undefined | Record<string, string | RegExp>} */
+                /** @type {undefined | Record<string, string>} */
                 let errorMessageData;
                 /** @type {null | SourceLocation} */
                 let errorMessageLoc = null;
@@ -921,7 +921,7 @@ const headerRule = {
                         if (!match(leadingLine, headerLine)) {
                             errorMessageId = "incorrectHeaderLine";
                             errorMessageData = {
-                                pattern: headerLine
+                                pattern: headerLine.toString()
                             };
                             const columnOffset = i === 0 ? "/*".length : 0;
                             assertDefined(firstLeadingCommentLoc);
